requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
PyMuPDF>=1.23.0
python-dotenv>=1.0.0
streamlit-pdf-viewer>=0.0.1 